% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParticleFilter.R
\name{particleFilter}
\alias{particleFilter}
\title{Run the naive particle filter}
\usage{
particleFilter(x, y, n)
}
\arguments{
\item{x}{Distribution of random variable (fix to be more specific)}

\item{y}{Observed data}

\item{n}{Number of particles simulated}
}
\value{
De-noised observations
}
\description{
Run the naive particle filter
}
\examples{
x_true = sim_data(50,0.4,1.2)$x
obs = sim_data(50,0.4,1.2)$y
estimates_means = particleFilter(x_true,obs,1000)
plot(x =1:50,y=estimates_means,type="l",col="red")
lines(obs,col="blue")
}
