% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particleFilterPredcit.R
\name{particleFilterPredcit}
\alias{particleFilterPredcit}
\title{particle filter prediction}
\usage{
particleFilterPredcit(x, y, n, period = 1, pred.times = 10, filter = TRUE)
}
\arguments{
\item{x}{Distribution of random variable (fix to be more specific)}

\item{y}{Observed data}

\item{n}{Number of particles simulated}

\item{period}{The period using in time series prediction}

\item{pred.times}{The prediction range}

\item{filter}{Whether use particle filter or not}
}
\value{
De-noised observations
}
\description{
particle filter prediction
}
\examples{
temp = sim_data(50,0.4,1.2)
particleFilter.Predcit(temp$x,temp$y,100,filter = TRUE)
particleFilter.Predcit(temp$x,temp$y,100,filter = FALSE)

}
