% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particleFilterPredict.R
\name{particleFilterPredict}
\alias{particleFilterPredict}
\title{particle filter prediction}
\usage{
particleFilterPredict(x, y, nParticle, period = 1, pred.n = 10, filter = TRUE)
}
\arguments{
\item{x}{The hidden state variables}

\item{y}{The observation variables}

\item{nParticle}{The number of particles simulated}

\item{period}{The period using in time series prediction}

\item{pred.n}{The prediction range}

\item{filter}{The option of whether use particle filter or not}
}
\value{
The prediction of observations
}
\description{
particle filter prediction
}
\examples{
temp = sim_data(50,0.4,1.2)
particleFilterPredict(temp$x,temp$y,100,filter = TRUE)
particleFilterPredict(temp$x,temp$y,100,filter = FALSE)
}
