% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_counts.R
\name{sim_data_counts}
\alias{sim_data_counts}
\title{Simulate the hidden variables X and observable variables Y for particle filters.
The observable variables (observation process) are related to
the hidden variables (state-process) by beta distribution in this simulation.
We will assume that in each state there are a finite number of "counts" of
successes and failures. The successes are modeled with a Poisson distribution
with rate "success_rate".}
\usage{
sim_data_counts(time, trials, success_rate, relation)
}
\arguments{
\item{time}{The number of observations for data simulated}

\item{trials}{The number of trials}

\item{success_rate}{The rate of success in the underlying process}

\item{relation}{The relation of X and Y}
}
\value{
A list of simulated X and Y vector, each with `length = time`
}
\description{
Simulate the hidden variables X and observable variables Y for particle filters.
The observable variables (observation process) are related to
the hidden variables (state-process) by beta distribution in this simulation.
We will assume that in each state there are a finite number of "counts" of
successes and failures. The successes are modeled with a Poisson distribution
with rate "success_rate".
}
\examples{
sims = sim_data(100,10,0.2, 0.2)
x_true = sims$x
obs = sims$y
}
