% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParticleFilter.R
\name{particleFilter}
\alias{particleFilter}
\title{The particle filter
(The sequential Monte Carlo with multinomial importance resampling (SIR))}
\usage{
particleFilter(x, y, nParticle)
}
\arguments{
\item{x}{The hidden state variables}

\item{y}{The observation variables}

\item{nParticle}{The number of particles simulated}
}
\value{
The de-noised observations
}
\description{
The particle filter
(The sequential Monte Carlo with multinomial importance resampling (SIR))
}
\examples{
x_true = sim_data(50,0.4,1.2)$x
obs = sim_data(50,0.4,1.2)$y
estimates_means = particleFilter(x_true,obs,1000)
plot(x =1:50,y=estimates_means,type="l",col="red")
lines(obs,col="blue")
}
