% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate X and Y data}
\usage{
sim_data(time, sx, sy)
}
\arguments{
\item{time}{Number of observations for data simulated}

\item{xs}{Variance of simulated random samples}

\item{ys}{Variance of simulated observations}
}
\value{
A list of simulated X and Y vector, each with `length = time`
}
\description{
Simulate X and Y data
}
\examples{
x_true = sim_data(50,0.4,1.2)$x
obs = sim_data(50,0.4,1.2)$y
}
